Move Pandas helper classes to their own files
